project_django/medium_clone/users/views.py
from rest_framework import generics
from django.contrib.auth import get_user_model
from .serializers import RegisterSerializer

# Create your views here.

class RegisterView(generics.CreateAPIView):
    queryset = get_user_model().objects.all()
    serializer_class = RegisterSerializer

    