project_django/requirements.txt
django
djangorestframework
psutil
djangorestframework-simplejwt
drf-yasg
redis
celery
django-celery-results